+++ src/main/kotlin/com/tezos/lang/michelson/editor/completion/MichelsonPatterns.kt
@@ -5,6 +5,8 @@ import com.tezos.lang.michelson.MichelsonTypes
 import com.tezos.lang.michelson.psi.MichelsonPsiFile
 import com.tezos.lang.michelson.psi.PsiBlockInstruction
 
+val PATTERN_IN_MICHELSON_FILE = PlatformPatterns.psiElement().inside(PlatformPatterns.psiFile(MichelsonPsiFile::class.java))!!
+
 val PATTERN_INSTRUCTION_TOKEN = PlatformPatterns.psiElement(MichelsonTypes.INSTRUCTION_TOKEN)!!
 
 val PATTERN_WITH_BLOCK_PARENT = PlatformPatterns.psiElement().withParent(PlatformPatterns.psiElement(PsiBlockInstruction::class.java))!!